# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moling API",
    "description": "墨灵 — AI 小说创作平台后端服务",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/secrets": {
      "get": {
        "tags": [
          "secrets"
        ],
        "summary": "List Secrets",
        "description": "List all secrets for a project.",
        "operationId": "list_secrets_api_v1_projects__project_id__secrets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SecretResp"
                  },
                  "title": "Response List Secrets Api V1 Projects  Project Id  Secrets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/secrets/character/{character_name}": {
      "get": {
        "tags": [
          "secrets"
        ],
        "summary": "Get Secrets By Character",
        "description": "Get secrets known to and unknown by a character (query by name).",
        "operationId": "get_secrets_by_character_api_v1_projects__project_id__secrets_character__character_name__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "character_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Secrets By Character Api V1 Projects  Project Id  Secrets Character  Character Name  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/secrets/character/{character_id}": {
      "put": {
        "tags": [
          "secrets"
        ],
        "summary": "Update Secrets By Character",
        "description": "Update secrets for a character (by ID, for backward compatibility).",
        "operationId": "update_secrets_by_character_api_v1_projects__project_id__secrets_character__character_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Character Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSecretsByCharacterReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Secrets By Character Api V1 Projects  Project Id  Secrets Character  Character Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/secrets/{secret_id}": {
      "patch": {
        "tags": [
          "secrets"
        ],
        "summary": "Update Secret",
        "description": "Update a specific secret (edit known_by, unknown_to, debt_count).",
        "operationId": "update_secret_api_v1_projects__project_id__secrets__secret_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "secret_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Secret Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSecretReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Secret Api V1 Projects  Project Id  Secrets  Secret Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SecretItemUpdate": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "related_characters": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Related Characters"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "secrecy_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secrecy Level"
          }
        },
        "type": "object",
        "title": "SecretItemUpdate",
        "description": "A secret item in the update request."
      },
      "SecretResp": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "known_by": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Known By"
          },
          "unknown_to": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unknown To"
          },
          "secrecy_level": {
            "type": "string",
            "title": "Secrecy Level"
          },
          "created_chapter": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created Chapter"
          },
          "debt": {
            "type": "integer",
            "title": "Debt"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "description",
          "known_by",
          "unknown_to",
          "secrecy_level",
          "debt",
          "created_at",
          "updated_at"
        ],
        "title": "SecretResp",
        "description": "Secret detail response."
      },
...
      "UpdateSecretReq": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "known_by": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Known By"
          },
          "unknown_to": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unknown To"
          },
          "secrecy_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secrecy Level"
          },
          "debt": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debt"
          }
        },
        "type": "object",
        "title": "UpdateSecretReq",
        "description": "Request body for updating a secret."
      },
      "UpdateSecretsByCharacterReq": {
        "properties": {
          "character_id": {
            "type": "integer",
            "title": "Character Id"
          },
          "character_name": {
            "type": "string",
            "title": "Character Name"
          },
          "secrets": {
            "items": {
              "$ref": "#/components/schemas/SecretItemUpdate"
            },
            "type": "array",
            "title": "Secrets"
          }
        },
        "type": "object",
        "required": [
          "character_id",
          "character_name",
          "secrets"
        ],
        "title": "UpdateSecretsByCharacterReq",
        "description": "Request body for updating secrets by character."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}